requirements.txt
openai>=1.40.0
gradio>=4.44.0
python-dotenv>=1.0.1
pypdf>=4.2.0
fpdf2>=2.7.9
